Breakout laika directory handling
